# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Real Estate CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: projectId
          in: query
          schema:
            type: string
        - name: assignedTo
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Leads list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
...
  /leads/{leadId}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get lead by ID
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete lead
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /leads/{leadId}/assign:
    post:
      operationId: assignLead
      tags: [leads]
      summary: Assign lead to sales
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignInput"
      responses:
        "200":
          description: Assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{leadId}/status:
    patch:
      operationId: updateLeadStatus
      tags: [leads]
      summary: Update lead status
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{leadId}/activities:
    get:
      operationId: listLeadActivities
      tags: [leads]
      summary: Get lead activities
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeadActivity"
    post:
      operationId: createLeadActivity
      tags: [leads]
      summary: Log activity on lead
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadActivity"

  # ── PROJECTS ──────────────────────────────────────────────────────────────
...
  /planner:
    get:
      operationId: listPlannerTasks
      tags: [planner]
      summary: List planner tasks
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlannerTask"
    post:
      operationId: createPlannerTask
      tags: [planner]
      summary: Create planner task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlannerTaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerTask"
  /planner/{taskId}:
    patch:
      operationId: updatePlannerTask
      tags: [planner]
      summary: Update planner task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlannerTaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerTask"
    delete:
      operationId: deletePlannerTask
      tags: [planner]
      summary: Delete planner task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── NOTIFICATIONS ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Lead:
      type: object
      required: [id, name, status, source, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        source:
          type: string
          enum: [manual, import, campaign, referral, website, social]
        status:
          type: string
          enum: [new, called, qualified, proposal, negotiation, won, lost]
        projectId:
          type: ["string", "null"]
        projectName:
          type: ["string", "null"]
        primarySalesId:
          type: ["string", "null"]
        primarySalesName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        nextAction:
          type: ["string", "null"]
        nextActionAt:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
        lastActionAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    LeadInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        source:
          type: string
          enum: [manual, import, campaign, referral, website, social]
        status:
          type: string
          enum: [new, called, qualified, proposal, negotiation, won, lost]
        projectId:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
    LeadUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        source:
          type: string
          enum: [manual, import, campaign, referral, website, social]
        status:
          type: string
          enum: [new, called, qualified, proposal, negotiation, won, lost]
        projectId:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        nextAction:
          type: ["string", "null"]
        nextActionAt:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
    LeadStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [new, called, qualified, proposal, negotiation, won, lost]
    AssignInput:
      type: object
      required: [salesId]
      properties:
        salesId:
          type: string
...
    LeadActivity:
      type: object
      required: [id, leadId, userId, type, createdAt]
      properties:
        id:
          type: string
        leadId:
          type: string
        userId:
          type: string
        userName:
          type: ["string", "null"]
        type:
          type: string
          enum: [call, message, meeting, email, note, status_change]
        notes:
          type: ["string", "null"]
        outcome:
          type: ["string", "null"]
        nextAction:
          type: ["string", "null"]
        nextActionAt:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        createdAt:
          type: string
    ActivityInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [call, message, meeting, email, note, status_change]
        notes:
          type: ["string", "null"]
        outcome:
          type: ["string", "null"]
        nextAction:
          type: ["string", "null"]
        nextActionAt:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]

    # ── PROJECT ────────────────────────────────────────────────────────────────
...
    PlannerTask:
      type: object
      required: [id, userId, date, title, isDone, priority, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        date:
          type: string
        title:
          type: string
        notes:
          type: ["string", "null"]
        isDone:
          type: boolean
        priority:
          type: string
          enum: [low, medium, high]
        position:
          type: integer
        createdAt:
          type: string
    PlannerTaskInput:
      type: object
      required: [date, title]
      properties:
        date:
          type: string
        title:
          type: string
          minLength: 1
        notes:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high]
        userId:
          type: ["string", "null"]
    PlannerTaskUpdate:
      type: object
      properties:
        title:
          type: string
        notes:
          type: ["string", "null"]
        isDone:
          type: boolean
        priority:
          type: string
          enum: [low, medium, high]
        position:
          type: integer

    # ── NOTIFICATION ────────────────────────────────────────────────────────────
...